homework125

+++ module-1/11_Inheritance/lecture-student/java/src/main/java/com/techelevator/Bid.java
@@ -1,20 +1,24 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Bid {
 
 	private final String bidder;
-	private final int bidAmount;
+	private final BigDecimal bidAmount;
 
 	public Bid(String bidder, int bidAmount) {
 		this.bidder = bidder;
-		this.bidAmount = bidAmount;
+		this.bidAmount = BigDecimal.valueOf(bidAmount);
 	}
 
 	public String getBidder() {
+
 		return bidder;
 	}
 
 	public int getBidAmount() {
-		return bidAmount;
+
+		return bidAmount.intValue();
 	}
 }
